+++ Assets/AppCenter/Plugins/AppCenterSDK/Crashes/Blank/CrashesInternal.cs
@@ -25,12 +25,9 @@ namespace Microsoft.AppCenter.Unity.Crashes.Internal
         {
         }
 
-        public static void TrackException(RawType exception)
+        public static string TrackException(RawType exception, IDictionary<string, string> properties, ErrorAttachmentLog[] attachments)
         {
-        }
-
-        public static void TrackException(RawType exception, IDictionary<string, string> properties)
-        {
+            return "";
         }
 
         public static AppCenterTask<bool> HasReceivedMemoryWarningInLastSessionAsync()
@@ -80,6 +77,15 @@ namespace Microsoft.AppCenter.Unity.Crashes.Internal
         }
 
         public static void StartCrashes()
+        {
+        }
+
+        public static ErrorReport BuildHandledErrorReport(string errorReportId)
+        {
+            return null;
+        }
+
+        public static void SendErrorAttachments(string errorReportId, ErrorAttachmentLog[] attachments)
         {
         }
     }